extracted timeline item into its own component
