+++ frontend/components/Domain/Recipe/RecipeTimeline.vue
@@ -7,87 +7,16 @@
     class="px-1"
     :style="maxHeight ? 'overflow-y: auto;' : ''"
   >
-    <v-timeline :dense="attrs.timeline.dense" class="timeline">
-      <v-timeline-item
+    <v-timeline :dense="$vuetify.breakpoint.smAndDown" class="timeline">
+      <RecipeTimelineItem
         v-for="(event, index) in timelineEvents"
         :key="event.id"
-        :class="attrs.timeline.item.class"
-        fill-dot
-        :small="attrs.timeline.item.small"
-        :icon="chooseEventIcon(event)"
-      >
-        <template v-if="!useMobileFormat" #opposite>
-          <v-chip v-if="event.timestamp" label large>
-            <v-icon class="mr-1"> {{ $globals.icons.calendar }} </v-icon>
-            {{ new Date(event.timestamp+"Z").toLocaleDateString($i18n.locale) }}
-          </v-chip>
-        </template>
-        <v-card>
-          <v-sheet>
-            <v-card-title>
-              <v-row>
-                <v-col align-self="center" :cols="useMobileFormat ? 'auto' : '2'" :class="attrs.timeline.item.avatar.class">
-                  <UserAvatar :user-id="event.userId" :size="attrs.timeline.item.avatar.size" />
-                </v-col>
-                <v-col v-if="useMobileFormat" align-self="center" class="pr-0">
-                  <v-chip label>
-                    <v-icon> {{ $globals.icons.calendar }} </v-icon>
-                    {{ new Date(event.timestamp+"Z").toLocaleDateString($i18n.locale) }}
-                  </v-chip>
-                </v-col>
-                <v-col v-else cols="9" style="margin: auto; text-align: center;">
-                  {{ event.subject }}
-                </v-col>
-                <v-spacer />
-                <v-col :cols="useMobileFormat ? 'auto' : '1'" class="px-0">
-                  <RecipeTimelineContextMenu
-                    v-if="$auth.user && $auth.user.id == event.userId && event.eventType != 'system'"
-                    :menu-top="false"
-                    :event="event"
-                    :menu-icon="$globals.icons.dotsVertical"
-                    fab
-                    color="transparent"
-                    :elevation="0"
-                    :card-menu="false"
-                    :use-items="{
-                      edit: true,
-                      delete: true,
-                    }"
-                    @update="updateTimelineEvent(index)"
-                    @delete="deleteTimelineEvent(index)"
-                  />
-                </v-col>
-              </v-row>
-            </v-card-title>
-            <v-sheet v-if="showRecipeCards && recipes.get(event.recipeId)">
-              <v-row class="pt-3 pb-7 mx-3" style="max-width: 100%;">
-                <v-col align-self="center" class="pa-0">
-                  <RecipeCardMobile
-                    :vertical="useMobileFormat"
-                    :name="recipes.get(event.recipeId).name"
-                    :slug="recipes.get(event.recipeId).slug"
-                    :description="recipes.get(event.recipeId).description"
-                    :rating="recipes.get(event.recipeId).rating"
-                    :image="recipes.get(event.recipeId).image"
-                    :recipe-id="recipes.get(event.recipeId).id"
-                  />
-                </v-col>
-              </v-row>
-            </v-sheet>
-            <v-divider v-if="showRecipeCards && recipes.get(event.recipeId) && (useMobileFormat || event.eventMessage)" />
-            <v-card-text>
-              <v-row>
-                <v-col>
-                  <strong v-if="useMobileFormat">{{ event.subject }}</strong>
-                  <div v-if="event.eventMessage" :class="useMobileFormat ? 'text-caption' : ''">
-                    {{ event.eventMessage }}
-                  </div>
-                </v-col>
-              </v-row>
-            </v-card-text>
-          </v-sheet>
-        </v-card>
-      </v-timeline-item>
+        :event="event"
+        :recipe="recipes.get(event.recipeId)"
+        :show-recipe-cards="showRecipeCards"
+        @update="updateTimelineEvent(index)"
+        @delete="deleteTimelineEvent(index)"
+      />
     </v-timeline>
   </v-card>
   <v-card v-else-if="!showRecipeCards">
@@ -98,17 +27,15 @@
 </template>
 
 <script lang="ts">
-import { computed, defineComponent, ref, useContext } from "@nuxtjs/composition-api";
+import { defineComponent, ref, useContext } from "@nuxtjs/composition-api";
 import { whenever } from "@vueuse/core";
-import RecipeCardMobile from "./RecipeCardMobile.vue";
-import RecipeTimelineContextMenu from "./RecipeTimelineContextMenu.vue";
+import RecipeTimelineItem from "./RecipeTimelineItem.vue"
 import { alert } from "~/composables/use-toast";
 import { useUserApi } from "~/composables/api";
 import { Recipe, RecipeTimelineEventOut, RecipeTimelineEventUpdate } from "~/lib/api/types/recipe"
-import UserAvatar from "~/components/Domain/User/UserAvatar.vue";
 
 export default defineComponent({
-  components: { RecipeCardMobile, RecipeTimelineContextMenu, UserAvatar },
+  components: { RecipeTimelineItem },
 
   props: {
     value: {
@@ -129,58 +56,11 @@ export default defineComponent({
     }
   },
 
-  setup(props, context) {
+  setup(props) {
     const api = useUserApi();
-    const { $globals, $vuetify, i18n } = useContext();
+    const { i18n } = useContext();
     const timelineEvents = ref([] as RecipeTimelineEventOut[]);
     const recipes = new Map<string, Recipe>();
-
-    const useMobileFormat = computed(() => {
-      return $vuetify.breakpoint.smAndDown;
-    });
-
-    const attrs = computed(() => {
-      if (useMobileFormat.value) {
-        return {
-          timeline: {
-            dense: true,
-            item: {
-              class: "px-0",
-              small: false,
-              avatar: {
-                size: "30px",
-                class: "pr-0",
-              },
-            }
-          }
-        }
-      }
-      else {
-        return {
-          timeline: {
-            dense: false,
-            item: {
-              class: "px-3",
-              small: false,
-              avatar: {
-                size: "42px",
-                class: "",
-              },
-            }
-          }
-        }
-      }
-    })
-
-    // V-Model Support
-    const dialog = computed({
-      get: () => {
-        return props.value;
-      },
-      set: (val) => {
-        context.emit("input", val);
-      },
-    });
 
     whenever(
       () => props.value,
@@ -188,22 +68,6 @@ export default defineComponent({
         refreshTimelineEvents();
       }
     );
-
-    function chooseEventIcon(event: RecipeTimelineEventOut) {
-      switch (event.eventType) {
-        case "comment":
-          return $globals.icons.commentTextMultiple;
-
-        case "info":
-          return $globals.icons.informationVariant;
-
-        case "system":
-          return $globals.icons.cog;
-
-        default:
-          return $globals.icons.informationVariant;
-      };
-    };
 
     // Timeline Actions
     async function updateTimelineEvent(index: number) {
@@ -286,15 +150,11 @@ export default defineComponent({
     refreshTimelineEvents();
 
     return {
-      attrs,
-      chooseEventIcon,
       deleteTimelineEvent,
-      dialog,
       recipes,
       refreshTimelineEvents,
       timelineEvents,
       updateTimelineEvent,
-      useMobileFormat,
     };
   },
 });
